Issue 147: pass -Wall -Werror on Solaris.
o Kludge for gotten const in struct pam_conv definition.
o Typecast pid_t to int for %d format.

+++ shellinabox/shellinaboxd.c
@@ -1280,7 +1280,7 @@ int main(int argc, char * const argv[]) {
     printf("X-ShellInABox-Port: %d\r\n"
            "X-ShellInABox-Pid: %d\r\n"
            "Content-type: text/html; charset=utf-8\r\n\r\n",
-           port, getpid());
+           port, (int)getpid());
     UNUSED(cgiRootSize);
     printfUnchecked(cgiRootStart, port, cgiSessionKey);
     fflush(stdout);
